+++ src/Dangl.SchneidControl/Models/Controllers/HeatMeterReplacements/HeatMeterReplacementViewModel.cs
@@ -1,0 +1,10 @@
+﻿namespace Dangl.SchneidControl.Models.Controllers.HeatMeterReplacements
+{
+    public class HeatMeterReplacementViewModel
+    {
+        public long Id { get; set; }
+        public DateTime ReplacedAtUtc { get; set; }
+        public int InitialValue { get; set; }
+        public int OldMeterValue { get; set; }
+    }
+}