+++ src/dangl-schneid-control/src/app/components/consumption/consumption.component.ts
@@ -1,21 +1,28 @@
-import * as moment from 'moment';
+import moment from 'moment';
 
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 import {
   ConsumptionClient,
   ConsumptionResolution,
 } from '../../generated-client/generated-client';
 
-import { MatDialogRef } from '@angular/material/dialog';
 import { Moment } from 'moment';
 import { Subscription } from 'rxjs';
+import { MatFormField, MatSuffix, MatLabel, MatSelect, MatOption } from '@angular/material/select';
+import { MatInput } from '@angular/material/input';
+import { MatDatepickerInput, MatDatepickerToggle, MatDatepicker } from '@angular/material/datepicker';
+import { FormsModule } from '@angular/forms';
+import { AreaChartModule } from '@swimlane/ngx-charts';
 
 @Component({
-  selector: 'app-stats',
-  templateUrl: './consumption.component.html',
-  styleUrls: ['./consumption.component.scss'],
+    selector: 'app-stats',
+    templateUrl: './consumption.component.html',
+    styleUrls: ['./consumption.component.scss'],
+    imports: [MatFormField, MatInput, MatDatepickerInput, FormsModule, MatDatepickerToggle, MatSuffix, MatDatepicker, MatLabel, MatSelect, MatOption, AreaChartModule]
 })
 export class ConsumptionComponent implements OnInit {
+  private consumptionClient = inject(ConsumptionClient);
+
   colorScheme = {
     domain: ['#00acc1'],
   };
@@ -55,11 +62,6 @@ export class ConsumptionComponent implements OnInit {
   get resolution() {
     return this._resolution;
   }
-
-  constructor(
-    private matDialog: MatDialogRef<ConsumptionComponent>,
-    private consumptionClient: ConsumptionClient
-  ) {}
 
   ngOnInit(): void {
     this.queryData();