+++ src/dangl-schneid-control/src/app/pipes/heating-circuit-status.pipe.ts
@@ -2,9 +2,7 @@ import { Pipe, PipeTransform } from '@angular/core';
 
 import { HeatingCircuitStatus } from '../generated-client/generated-client';
 
-@Pipe({
-  name: 'heatingCircuitStatus',
-})
+@Pipe({ name: 'heatingCircuitStatus' })
 export class HeatingCircuitStatusPipe implements PipeTransform {
   transform(value?: HeatingCircuitStatus): string {
     if (value) {